refactor: extract useLayoutCustomDrag with profile and action refs
Made-with: Cursor
